Refactor allow-related apis into allowed_functions.py
Seems generally good, but also helps make allow* apis usable
from other dynamo modules (impossible to import allow_in_graph
in some places previously due to circular import)

[ghstack-poisoned]
